feat: Add portable Image component
