+++ apps/next/app/NextClientRootLayout.tsx
@@ -1,6 +1,7 @@
 'use client'
 import React from 'react'
 import UniversalAppProviders from '@app/core/screens/UniversalAppProviders'
+import { Image as NextContextImage } from '@app/core/components/Image.next'
 import { Link as NextContextLink } from '@app/core/navigation/Link.next'
 import { useRouter as useNextContextRouter } from '@app/core/navigation/useRouter.next'
 import { useRouteParams as useNextRouteParams } from '@app/core/navigation/useRouteParams.next'
@@ -26,6 +27,7 @@ const NextClientRootLayout = ({ children }: NextClientRootLayoutProps) => {
     
     return (
         <UniversalAppProviders
+            contextImage={NextContextImage}
             contextLink={NextContextLink}
             contextRouter={nextContextRouter}
             useContextRouteParams={useNextRouteParams}
